src/requirements.txt
socket
ssl
os
subprocess
pyjwt
asyncpg
python-dotenv